chore(faz0): clippy --fix mekanik temizlik (130→86 uyarı, davranış değişmedi)
Modernizasyon Faz 0 — güvenlik ağı. cargo clippy --fix ile makine-uygulanabilir
44 düzeltme (redundant closure, manual Default/impl, useless conversion,
first()/get(0), match-scrutinee vb.) 62 dosyada uygulandı. Kalan 86 uyarı
elle yargı gerektirenler (too-many-arguments, manuel Default) — sonraki
fazların bilinçli işi.

Doğrulama: 357/357 lib test yeşil, release derleme temiz (29.6s).

+++ memos_trading_core/src/core/bridge.rs
@@ -213,7 +213,7 @@ pub fn get_snapshot(st: &AppState) -> MissionControl {
         drift_score: hub_drift,
         mc_ruin_prob: 0.01,
         is_evolution_active: hub_evolution_active,
-        next_evolution_secs: 300_u64.saturating_sub((hub_cycle % 300) as u64),
+        next_evolution_secs: 300_u64.saturating_sub(hub_cycle % 300),
         live_strategy: live_strategy_name,
         controller_state: hub_state,
         controller_cycle: hub_cycle,
